backend/pyproject.toml
[project]
name = "iluminacasa"
version = "0.1.0"
description = "IluminaCasa API"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=4.3.0",
    "fastapi[standard]>=0.116.1",
    "httpx>=0.28.1",
    "passlib>=1.7.4",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "slowapi>=0.1.9",
    "sqlmodel>=0.0.24",
]
